# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BE API — WP-02 Identidad y sesiones · WP-03 Vínculo, consentimiento y PDP",
    "version": "0.3.0",
    "description": "Generado desde @be/domain (contratos.ts y contratos-vinculo.ts). No editar a mano."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "sesion": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Sesión revocable verificada en cada request (08 §26)."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/relationships/{relationshipId}/consent-requirements": {
      "get": {
        "operationId": "API-CON-01",
        "summary": "Texto y versión de B2 aplicables a un vínculo propio (solo el asesorado titular).",
        "description": "Fuente: 09v8:1500-1566 · DEUDA_LEGAJO DL-038/DL-039. AuthN: SESSION.",
        "security": [
          {
            "sesion": []
          }
        ],
        "parameters": [
          {
            "name": "relationshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-BE-Surface",
            "in": "header",
            "required": false,
            "description": "Superficie declarada (procedencia). Nunca autoriza (DEUDA_LEGAJO DL-022).",
            "schema": {
              "type": "string",
              "enum": [
                "WEB",
                "APK"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "relationshipId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "professional": {
                          "type": "object",
                          "properties": {
                            "identityId": {
                              "type": "string",
                              "minLength": 1
                            },
                            "displayName": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "identityId",
                            "displayName"
                          ],
                          "additionalProperties": false
                        },
                        "scope": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "NUTRICION",
                                "ENTRENAMIENTO",
                                "ANTROPOMETRIA"
                              ]
                            },
                            "label": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "code",
                            "label"
                          ],
                          "additionalProperties": false
                        },
                        "purpose": {
                          "type": "string",
                          "enum": [
                            "ACOMPANAMIENTO_NUTRICIONAL",
                            "PLANIFICACION_DEL_ENTRENAMIENTO",
                            "EVALUACION_ANTROPOMETRICA"
                          ]
                        },
                        "consentVersion": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "minLength": 1
                            },
                            "text": {
                              "type": "string"
                            },
                            "textHash": {
                              "type": "string"
                            },
                            "effectiveFrom": {
                              "type": "string",
                              "format": "date-time",
                              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                            }
                          },
                          "required": [
                            "id",
                            "text",
                            "textHash",
                            "effectiveFrom"
                          ],
                          "additionalProperties": false
                        },
                        "pertinentCategories": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "label": {
                                "type": "string"
                              },
                              "detailLevel": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "label",
                              "detailLevel"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "professionalProfileDisclosure": {
                          "type": "object",
                          "properties": {
                            "profileType": {
                              "type": "string",
                              "enum": [
                                "HEALTH_PROFESSIONAL",
                                "NON_HEALTH_PROFESSIONAL"
                              ]
                            },
                            "notice": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "profileType",
                            "notice"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "relationshipId",
                        "professional",
                        "scope",
                        "purpose",
                        "consentVersion",
                        "pertinentCategories",
                        "professionalProfileDisclosure"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "ErrorEnvelope: INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "ErrorEnvelope: AUTHENTICATION_REQUIRED | SESSION_INVALID | SESSION_EXPIRED | SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "ErrorEnvelope: RESOURCE_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "ErrorEnvelope: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "ErrorEnvelope: DB_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/relationships/{relationshipId}/consents": {
      "post": {
        "operationId": "API-CON-02",
        "summary": "Otorgar B2 para la versión presentada (solo el asesorado titular). 201 al otorgar; 200 si ya estaba vigente con esa versión o al aceptar una versión nueva o reotorgar.",
        "description": "Fuente: 09v8:1570-1660 · DEUDA_LEGAJO DL-038. AuthN: SESSION.",
        "security": [
          {
            "sesion": []
          }
        ],
        "parameters": [
          {
            "name": "relationshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{8,128}$"
            }
          },
          {
            "name": "X-BE-Surface",
            "in": "header",
            "required": false,
            "description": "Superficie declarada (procedencia). Nunca autoriza (DEUDA_LEGAJO DL-022).",
            "schema": {
              "type": "string",
              "enum": [
                "WEB",
                "APK"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "consentVersionId": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "consentVersionId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "consentId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "relationshipId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "consentVersionId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "state": {
                          "type": "string",
                          "const": "ACTIVE"
                        },
                        "acceptedAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                        },
                        "scope": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "NUTRICION",
                                "ENTRENAMIENTO",
                                "ANTROPOMETRIA"
                              ]
                            },
                            "label": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "code",
                            "label"
                          ],
                          "additionalProperties": false
                        },
                        "purpose": {
                          "type": "string",
                          "enum": [
                            "ACOMPANAMIENTO_NUTRICIONAL",
                            "PLANIFICACION_DEL_ENTRENAMIENTO",
                            "EVALUACION_ANTROPOMETRICA"
                          ]
                        },
                        "pertinentCategories": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "label": {
                                "type": "string"
                              },
                              "detailLevel": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "label",
                              "detailLevel"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "consentId",
                        "relationshipId",
                        "consentVersionId",
                        "state",
                        "acceptedAt",
                        "scope",
                        "purpose",
                        "pertinentCategories"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "Éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "consentId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "relationshipId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "consentVersionId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "state": {
                          "type": "string",
                          "const": "ACTIVE"
                        },
                        "acceptedAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                        },
                        "scope": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "NUTRICION",
                                "ENTRENAMIENTO",
                                "ANTROPOMETRIA"
                              ]
                            },
                            "label": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "code",
                            "label"
                          ],
                          "additionalProperties": false
                        },
                        "purpose": {
                          "type": "string",
                          "enum": [
                            "ACOMPANAMIENTO_NUTRICIONAL",
                            "PLANIFICACION_DEL_ENTRENAMIENTO",
                            "EVALUACION_ANTROPOMETRICA"
                          ]
                        },
                        "pertinentCategories": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "label": {
                                "type": "string"
                              },
                              "detailLevel": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "label",
                              "detailLevel"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "consentId",
                        "relationshipId",
                        "consentVersionId",
                        "state",
                        "acceptedAt",
                        "scope",
                        "purpose",
                        "pertinentCategories"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "ErrorEnvelope: INVALID_REQUEST | UNKNOWN_FIELD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "ErrorEnvelope: AUTHENTICATION_REQUIRED | SESSION_INVALID | SESSION_EXPIRED | SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "ErrorEnvelope: RESOURCE_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "ErrorEnvelope: CONSENT_VERSION_STALE | IDEMPOTENCY_KEY_REUSED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "ErrorEnvelope: RELATIONSHIP_NOT_READY_FOR_CONSENT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "ErrorEnvelope: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "ErrorEnvelope: DB_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/consents/{consentId}/revoke": {
      "post": {
        "operationId": "API-CON-04",
        "summary": "Revocar B2. Idempotente por semántica (sin Idempotency-Key). Corta el acceso futuro; no finaliza el vínculo.",
        "description": "Fuente: 09v8:1699-1758. AuthN: SESSION.",
        "security": [
          {
            "sesion": []
          }
        ],
        "parameters": [
          {
            "name": "consentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-BE-Surface",
            "in": "header",
            "required": false,
            "description": "Superficie declarada (procedencia). Nunca autoriza (DEUDA_LEGAJO DL-022).",
            "schema": {
              "type": "string",
              "enum": [
                "WEB",
                "APK"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "consentId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "state": {
                          "type": "string",
                          "const": "REVOKED"
                        },
                        "revokedAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                        }
                      },
                      "required": [
                        "consentId",
                        "state",
                        "revokedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "ErrorEnvelope: INVALID_REQUEST | UNKNOWN_FIELD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "ErrorEnvelope: AUTHENTICATION_REQUIRED | SESSION_INVALID | SESSION_EXPIRED | SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "ErrorEnvelope: RESOURCE_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "ErrorEnvelope: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "ErrorEnvelope: DB_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}